feat: support extended graphql types

+++ src/main/java/com/example/demo/author/Author.java
@@ -1,5 +1,3 @@
 package com.example.demo.author;
 
-import java.math.BigInteger;
-
-public record Author(BigInteger id, String firstName, String lastName) {}
+public record Author(Long id, String firstName, String lastName) {}
